Feat: Result feedback structure implemented

+++ Entities/DTOs/CarDetailDTO.cs
@@ -8,6 +8,6 @@ namespace Entities.DTOs
         public string CarName { get; set; }
         public string BrandName { get; set; }
         public string ColorName { get; set; }
-        public int DailyPrice { get; set; }
+        public double DailyPrice { get; set; }
     }
 }
